+++ changelog.md
@@ -3,6 +3,7 @@
 ### Bug Fixes
 
 * Application in production was not finding custom OpenBCI logo.
+* Process command error would send message type error to GUI for all boards. Changes to send command type with error code.
 
 # v1.3.5